requirements.txt
decorator==4.0.11
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
numpy==1.16.1
Pillow==5.4.1
requests==2.11.1
requests-toolbelt==0.7.0
tqdm==4.11.2
